# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bilingual personal website + CMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pages/{pageId}/sections:
    get:
      operationId: listSections
      tags: [sections]
      summary: List all sections for a page
      parameters:
        - in: path
          name: pageId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of sections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SectionWithImages"
    post:
      operationId: createSection
      tags: [sections]
      summary: Add a section to a page
      parameters:
        - in: path
          name: pageId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionInput"
      responses:
        "201":
          description: Section created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SectionWithImages"
  /sections/{id}:
    get:
      operationId: getSection
      tags: [sections]
      summary: Get a section with its images
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Section with images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SectionWithImages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateSection
      tags: [sections]
      summary: Update a section's data
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SectionUpdate"
      responses:
        "200":
          description: Updated section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SectionWithImages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteSection
      tags: [sections]
      summary: Delete a section
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /sections/{id}/visibility:
    patch:
      operationId: toggleSectionVisibility
      tags: [sections]
      summary: Toggle section visibility
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisibilityUpdate"
      responses:
        "200":
          description: Updated section
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SectionWithImages"
  /pages/{pageId}/sections/reorder:
    post:
      operationId: reorderSections
      tags: [sections]
      summary: Reorder sections on a page
      parameters:
        - in: path
          name: pageId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderRequest"
      responses:
        "200":
          description: Reordered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── Images ────────────────────────────────────────────────────────────────
  # Note: POST /api/sections/:sectionId/images uses multipart/form-data handled
  # directly by the Express route with Multer — not in this spec to avoid
  # codegen complexity with binary file types.
...
  /public/pages/{slug}:
    get:
      operationId: getPublicPage
      tags: [public]
      summary: Get a published page by slug with visible sections
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public page with sections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPage"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth ──────────────────────────────────────────────────────────────
...
    SectionDataItem:
      type: object
      required: [id]
      properties:
        id:
          type: string
        titleAr:
          type: string
        titleEn:
          type: string
        descriptionAr:
          type: string
        descriptionEn:
          type: string
        subheadingAr:
          type: string
        subheadingEn:
          type: string
        date:
          type: string
        icon:
          type: string
        bullets:
          type: array
          items:
            $ref: "#/components/schemas/BulletItem"
    BulletItem:
      type: object
      required: [id]
      properties:
        id:
          type: string
        textAr:
          type: string
        textEn:
          type: string
    SectionData:
      type: object
      properties:
        titleAr:
          type: string
        titleEn:
          type: string
        contentAr:
          type: string
        contentEn:
          type: string
        imagePosition:
          type: string
        email:
          type: string
        linkedin:
          type: string
        locationAr:
          type: string
        locationEn:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SectionDataItem"

    # ── Sections ──────────────────────────────────────────────────────────
...
    SectionWithImages:
      type: object
      required: [id, pageId, type, sortOrder, isVisible, data, images]
      properties:
        id:
          type: integer
        pageId:
          type: integer
        type:
          type: string
          enum: [hero, text, text_with_image, image_gallery, cards_grid, timeline, contact_strip]
        sortOrder:
          type: integer
        isVisible:
          type: boolean
        data:
          $ref: "#/components/schemas/SectionData"
        images:
          type: array
          items:
            $ref: "#/components/schemas/SectionImage"
    SectionInput:
      type: object
      required: [type, data]
      properties:
        type:
          type: string
          enum: [hero, text, text_with_image, image_gallery, cards_grid, timeline, contact_strip]
        data:
          $ref: "#/components/schemas/SectionData"
    SectionUpdate:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/SectionData"
    VisibilityUpdate:
      type: object
      required: [isVisible]
      properties:
        isVisible:
          type: boolean
    ReorderRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: integer

    # ── Images ────────────────────────────────────────────────────────────
    SectionImage:
      type: object
      required: [id, sectionId, filename, captionAr, captionEn, sortOrder, url]
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        filename:
          type: string
        captionAr:
          type: string
        captionEn:
          type: string
        sortOrder:
          type: integer
        url:
          type: string
...
    PublicPage:
      type: object
      required: [id, titleAr, titleEn, slug, isHomepage, sections]
      properties:
        id:
          type: integer
        titleAr:
          type: string
        titleEn:
          type: string
        slug:
          type: string
        isHomepage:
          type: boolean
        sections:
          type: array
          items:
            $ref: "#/components/schemas/SectionWithImages"